# Back/docs/swagger.json
{
  "openapi": "3.1.0",
  "info": {
    "version": "1.0.0",
    "title": "TesteoramaBackendSwagger",
    "license": {
      "name": "Apache 2.0",
      "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
    },
    "description": "Backend for Tasteorama studying project."
  },
  "servers": [
    {
      "url": "http://localhost:3000"
    }
  ],
...
  "paths": {
...
    "/recipes/{recipeId}": {
      "get": {
        "tags": [
          "Recipes"
        ],
        "summary": "Get recipe by Id",
        "operationId": "getRecipeById",
        "description": "Get recipes by Id",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "recipeId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "example": "65ca67e7ae7f10c88b598384"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully found recipe!",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "status",
                    "message",
                    "data"
                  ],
                  "properties": {
                    "status": {
                      "type": "integer",
                      "example": 200
                    },
                    "message": {
                      "type": "string",
                      "example": "Successfully found recipe!"
                    },
                    "data": {
                      "$ref": "#/components/schemas/recipeDetailed"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "$ref": "#/components/responses/404"
          }
        }
      },
      "delete": {
        "tags": [
          "Recipes"
        ],
        "summary": "Delete recipe by id",
        "operationId": "deleteRecipe",
        "description": "Delete recipe by id",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "recipeId",
            "required": true,
            "schema": {
              "type": "string",
              "example": "65ca67e7ae7f10c88b598384"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Deleted recipe with id"
          },
          "401": {
            "$ref": "#/components/responses/401"
          },
          "404": {
            "$ref": "#/components/responses/404"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer"
      }
    },
    "schemas": {
...
      "ingredient": {
        "type": "object",
        "required": [
          "_id",
          "name"
        ],
        "properties": {
          "_id": {
            "description": "id of ingredient",
            "type": "string",
            "example": "640c2dd963a319ea671e37aa"
          },
          "name": {
            "description": "Name of ingredient",
            "type": "string",
            "example": "Squid"
          }
        }
      },
      "recipeDetailed": {
        "type": "object",
        "required": [
          "_id",
          "title",
          "category",
          "owner",
          "instructions",
          "description",
          "thumb",
          "time",
          "ingredients"
        ],
        "properties": {
          "_id": {
            "description": "Id of recipe",
            "type": "string",
            "example": "65e4decdd286b30065d54af9"
          },
          "title": {
            "description": "Title of a dish",
            "type": "string",
            "example": "Tuna and Egg Briks"
          },
          "category": {
            "description": "Name of category",
            "type": "string",
            "example": "Desert"
          },
          "owner": {
            "description": "Id of dish's creator",
            "type": "string",
            "example": "65e4decdd286b30065d54af9"
          },
          "instructions": {
            "description": "Steps for preparing dish",
            "type": "string",
            "example": "Heat 2 tsp of the oil in a large saucepan and cook the spring onions over a low heat for 3 minutes or until beginning to soften. Add the spinach, cover with a tight-fitting lid and cook for a further 2–3 minutes or until tender and wilted, stirring once or twice. Tip the mixture into a sieve or colander and leave to drain and cool. \\nUsing a saucer as a guide, cut out 24 rounds about 12.5 cm (5 in) in diameter from the filo pastry, cutting 6 rounds from each sheet. Stack the filo rounds in a pile, then cover with cling film to prevent them from drying out..."
          },
          "description": {
            "description": "Short description of dish",
            "type": "string",
            "example": "A crispy Tunisian pastry filled with tuna, egg, onion, and parsley, spiced with cumin and harissa."
          },
          "thumb": {
            "description": "Url to thumbnail to dish",
            "type": "string",
            "example": "https://ftp.goit.study/img/so-yummy/preview/Tuna%20and%20Egg%20Briks.jpg"
          },
          "img": {
            "description": "Url to image of a dish",
            "type": "string",
            "example": "https://ftp.goit.study/img/so-yummy/preview/Tuna%20and%20Egg%20Briks.jpg"
          },
          "time": {
            "description": "Time needed to prepare a dish",
            "type": "string",
            "example": "30 min"
          },
          "ingredients": {
            "description": "List of ingredients",
            "type": "array",
            "items": {
              "type": "object",
              "required": [
                "ingredient",
                "measure"
              ],
              "properties": {
                "ingredient": {
                  "$ref": "#/components/schemas/ingredient"
                },
                "measure": {
                  "description": "Measure of how much is needed",
                  "type": "string",
                  "example": "100ml"
                }
              }
            }
          }
        }
      },
...
    },
    "responses": {
      "401": {
        "description": "Unauthorized error",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "required": [
                "status",
                "message",
                "data"
              ],
              "properties": {
                "status": {
                  "type": "integer",
                  "example": 401
                },
                "message": {
                  "type": "string",
                  "example": "UnauthorizedError"
                },
                "data": {
                  "type": "object",
                  "required": [
                    "message"
                  ],
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Access token expired"
                    }
                  }
                }
              }
            }
          }
        }
      },
      "404": {
        "description": "Not found",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "required": [
                "status",
                "message",
                "data"
              ],
              "properties": {
                "status": {
                  "type": "integer",
                  "example": 404
                },
                "message": {
                  "type": "string",
                  "example": "Something* not found"
                },
                "data": {
                  "type": "object",
                  "required": [
                    "message"
                  ],
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Something* not found"
                    }
                  }
                }
              }
            }
          }
        }
      }
    }
  }
}